# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/event": {
            "get": {
                "description": "Retrieve events from ClickHouse.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Get"
                ],
                "summary": "Get Events",
                "responses": {
                    "200": {
                        "description": "events retrieved successfully",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controller.Event"
                            }
                        }
                    },
                    "500": {
                        "description": "error getting events from ClickHouse",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Insert event into ClickHouse.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Create"
                ],
                "summary": "Insert Event",
                "parameters": [
                    {
                        "description": "Event object to insert",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Event"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "event inserted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/search": {
            "get": {
                "description": "Search for events in ClickHouse based on criteria.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Get"
                ],
                "summary": "Search Events",
                "parameters": [
                    {
                        "description": "Search criteria",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "events matching search criteria",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controller.Event"
                            }
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error parsing events from ClickHouse",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/test": {
            "get": {
                "description": "Insert random test data.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Get"
                ],
                "summary": "Test",
                "responses": {
                    "200": {
                        "description": "test data inserted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.Event": {
            "type": "object",
            "properties": {
                "eventID": {
                    "type": "integer"
                },
                "eventTime": {
                    "type": "string"
                },
                "payload": {
                    "type": "integer"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "controller.Request": {
            "type": "object",
            "properties": {
                "eventID": {
                    "type": "integer"
                },
                "lower limit": {
                    "type": "string"
                },
                "upper limit": {
                    "type": "string"
                }
            }
        }
    }
}